chore: move mosip utils to own file and separate birth/death
